[FX] Move graph_manipulation and param_fetch out of experimental and into passes
Differential Revision: [D31007630](https://our.internmc.facebook.com/intern/diff/D31007630/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D31007630/)!

[ghstack-poisoned]
